Module Octez_smart_rollup_node_Proxford.Daemon_helpers

val check_pvm_initial_state_hash : _ Octez_smart_rollup_node.Node_context.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Ensure that the initial state hash of the PVM as defined by the rollup node matches the one of the PVM on the L1 node.

val process_l1_block_operations : catching_up:bool -> Octez_smart_rollup_node.Node_context.rw -> Octez_smart_rollup_node.Layer1.header -> unit Tezos_base.TzPervasives.tzresult Lwt.t

React to L1 operations included in a block of the chain.